account/urls.py
from django.urls import path
from rest_framework.urls import app_name
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView

from account.views import RegistrationAPIView

app_name = 'account'

urlpatterns = [
    path('', RegistrationAPIView.as_view(), name='registration'),
    path('token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
]