requirements.txt
asgiref==3.8.1
Django==5.1.7
django-ckeditor==6.7.2
django-filter==25.1
django-js-asset==3.1.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.9
inflection==0.5.1
Markdown==3.7
mysqlclient==2.2.7
packaging==24.2
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
